requirements.txt
# Core AI and Environment
google-genai>=1.73.1
python-dotenv>=1.2.2

# Web Interface
gradio>=6.12.0

# Search Fallback
tavily-python>=0.7.23

# Vector Database & Processing
pypdf>=6.10.2
faiss-cpu>=1.13.2
numpy>=2.4.4

# Testing
pytest>=9.0.3
